+++ OOPCourse.Main/Utilities/MoneyConverter.cs
@@ -2,7 +2,7 @@
 {
     internal class MoneyConverter
     {
-        public static string Convert(double reward)
+        public static string Convert(decimal reward)
         {
             if (reward < 1)
                 return $"{reward * 100} pennies";
